+++ tests/test-jpms/test-jpms-websocket-core/src/test/java/WebSocketCoreJPMSTest.java
@@ -17,18 +17,21 @@ import java.util.concurrent.TimeUnit;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.util.Callback;
+import org.eclipse.jetty.websocket.core.CloseStatus;
 import org.eclipse.jetty.websocket.core.CoreSession;
+import org.eclipse.jetty.websocket.core.Frame;
 import org.eclipse.jetty.websocket.core.FrameHandler;
 import org.eclipse.jetty.websocket.core.client.CoreClientUpgradeRequest;
 import org.eclipse.jetty.websocket.core.client.WebSocketCoreClient;
 import org.eclipse.jetty.websocket.core.server.WebSocketNegotiator;
 import org.eclipse.jetty.websocket.core.server.WebSocketUpgradeHandler;
-import org.example.websocket.MyFrameHandler;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public class WebSocketCoreJpmsTest
+public class WebSocketCoreJPMSTest
 {
     private Server _server;
     private ServerConnector _serverConnector;
@@ -42,7 +45,7 @@ public class WebSocketCoreJpmsTest
         _server.addConnector(_serverConnector);
 
         WebSocketUpgradeHandler webSocketUpgradeHandler = new WebSocketUpgradeHandler();
-        FrameHandler myFrameHandler = new MyFrameHandler("Server");
+        FrameHandler myFrameHandler = new TestFrameHandler("Server");
         webSocketUpgradeHandler.addMapping("/ws", WebSocketNegotiator.from(negotiation -> myFrameHandler));
 
         _server.setHandler(webSocketUpgradeHandler);
@@ -62,11 +65,51 @@ public class WebSocketCoreJpmsTest
     @Test
     public void testSimpleEcho() throws Exception
     {
-        MyFrameHandler frameHandler = new MyFrameHandler("Client");
+        TestFrameHandler frameHandler = new TestFrameHandler("Client");
         URI uri = URI.create("ws://localhost:" + _serverConnector.getLocalPort() + "/ws");
         CoreClientUpgradeRequest upgradeRequest = CoreClientUpgradeRequest.from(_client, uri, frameHandler);
         upgradeRequest.addExtensions("permessage-deflate");
         CoreSession coreSession = _client.connect(upgradeRequest).get(5, TimeUnit.SECONDS);
         coreSession.close(Callback.NOOP);
+    }
+
+    public static class TestFrameHandler implements FrameHandler
+    {
+        private static final Logger LOG = LoggerFactory.getLogger(TestFrameHandler.class);
+
+        private final String _id;
+
+        public TestFrameHandler(String id)
+        {
+            _id = id;
+        }
+
+        @Override
+        public void onOpen(CoreSession coreSession, Callback callback)
+        {
+            LOG.info(_id + " onOpen");
+            callback.succeeded();
+        }
+
+        @Override
+        public void onFrame(Frame frame, Callback callback)
+        {
+            LOG.info(_id + " onFrame");
+            callback.succeeded();
+        }
+
+        @Override
+        public void onError(Throwable cause, Callback callback)
+        {
+            LOG.info(_id + " onError");
+            callback.succeeded();
+        }
+
+        @Override
+        public void onClosed(CloseStatus closeStatus, Callback callback)
+        {
+            LOG.info(_id + " onClosed");
+            callback.succeeded();
+        }
     }
 }
